fix: Renamed admin files, fixed imports, and resolved button issues
