--- src/componentsAdmin/Footer.css
@@ -1,21 +1,0 @@
-.footer {
-  background-color: #2c3e50;
-  color: #ecf0f1;
-  padding: 1rem;
-  text-align: center;
-  margin-top: auto;
-}
-
-.footer-content {
-  max-width: 1200px;
-  margin: 0 auto;
-}
-
-.footer p {
-  margin: 0.5rem 0;
-}
-
-.version {
-  font-size: 0.8rem;
-  color: #bdc3c7;
-}
