--- src/componentsAdmin/Header.css
@@ -1,86 +1,0 @@
-.header {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 1rem 2rem;
-  background-color: #2c3e50;
-  color: white;
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-}
-
-.logo {
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-}
-
-.logo h1 {
-  margin: 0;
-  font-size: 2rem;
-  font-weight: 700;
-  color: #3498db;
-}
-
-.subtitle {
-  margin: 0;
-  font-size: 0.9rem;
-  color: #ecf0f1;
-}
-
-.nav ul {
-  display: flex;
-  list-style: none;
-  margin: 0;
-  padding: 0;
-}
-
-.nav li {
-  margin-left: 1.5rem;
-}
-
-.nav a {
-  color: #ecf0f1;
-  text-decoration: none;
-  font-weight: 500;
-  transition: color 0.3s;
-}
-
-.nav a:hover {
-  color: #3498db;
-}
-
-.nav-link svg {
-  margin-left: -10px;
-  margin-top: 0px;
- /* color: #fff;*/
-  transition: color 0.3s ease, transform 0.3s ease;
-}
-
-.nav-link:hover svg {
-  color: #3498db;
-  transform: scale(1.1);
-  cursor: pointer
-}
-
-@media (max-width: 768px) {
-  .header {
-    flex-direction: column;
-    padding: 1rem;
-  }
-  
-  .logo {
-    margin-bottom: 1rem;
-    align-items: center;
-  }
-  
-  .nav ul {
-    flex-direction: column;
-    align-items: center;
-  }
-  
-  .nav li {
-    margin: 0.5rem 0;
-    margin-left: 0;
-  }
- 
-}
