--- src/componentsAdmin/ReservationCard.css
@@ -1,94 +1,0 @@
-.reservation-card {
-  background-color: white;
-  border-radius: 8px;
-  box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
-  overflow: hidden;
-  margin-bottom: 1.5rem;
-}
-
-.reservation-header {
-  padding: 1.2rem 1.5rem;
-  background-color: #f8f9fa;
-  border-bottom: 1px solid #e9ecef;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.reservation-header h3 {
-  margin: 0;
-  color: #2c3e50;
-  font-size: 1.2rem;
-}
-
-.reservation-status {
-  padding: 0.3rem 0.8rem;
-  border-radius: 20px;
-  font-size: 0.8rem;
-  font-weight: 600;
-}
-
-.reservation-status.confirmada {
-  background-color: #e3f7ed;
-  color: #27ae60;
-}
-
-.reservation-status.pendente {
-  background-color: #fff8e1;
-  color: #f39c12;
-}
-
-.reservation-status.cancelada {
-  background-color: #fde8e8;
-  color: #e74c3c;
-}
-
-.reservation-details {
-  padding: 1.5rem;
-}
-
-.detail-item {
-  margin-bottom: 0.8rem;
-  display: flex;
-}
-
-.detail-label {
-  font-weight: 600;
-  color: #7f8c8d;
-  width: 120px;
-  flex-shrink: 0;
-}
-
-.detail-value {
-  color: #2c3e50;
-}
-
-.detail-value.purpose {
-  white-space: pre-line;
-}
-
-.cancel-reservation-button {
-  background-color: #e74c3c;
-  color: white;
-  border: none;
-  padding: 1rem;
-  width: 100%;
-  font-weight: 600;
-  cursor: pointer;
-  transition: background-color 0.3s;
-}
-
-.cancel-reservation-button:hover {
-  background-color: #c0392b;
-}
-
-@media (max-width: 768px) {
-  .detail-item {
-    flex-direction: column;
-  }
-  
-  .detail-label {
-    width: 100%;
-    margin-bottom: 0.2rem;
-  }
-}
