--- src/componentsAdmin/RoomCard.css
@@ -1,96 +1,0 @@
-.room-card {
-  background-color: white;
-  border-radius: 8px;
-  box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
-  overflow: hidden;
-  transition: transform 0.3s, box-shadow 0.3s;
-  display: flex;
-  flex-direction: column;
-  height: 100%;
-}
-
-.room-card:hover {
-  transform: translateY(-5px);
-  box-shadow: 0 5px 15px rgba(0, 0, 0, 0.2);
-}
-
-.room-image {
-  position: relative;
-  height: 200px;
-  overflow: hidden;
-}
-
-.room-image img {
-  width: 100%;
-  height: 100%;
-  object-fit: cover;
-}
-
-.room-capacity {
-  position: absolute;
-  bottom: 10px;
-  right: 10px;
-  background-color: rgba(44, 62, 80, 0.8);
-  color: white;
-  padding: 5px 10px;
-  border-radius: 20px;
-  font-size: 0.8rem;
-}
-
-.room-info {
-  padding: 1.5rem;
-  flex-grow: 1;
-}
-
-.room-info h3 {
-  margin: 0 0 0.5rem;
-  color: #2c3e50;
-  font-size: 1.3rem;
-}
-
-.room-location {
-  color: #7f8c8d;
-  font-size: 0.9rem;
-  margin-bottom: 1rem;
-}
-
-.room-description {
-  color: #34495e;
-  margin-bottom: 1rem;
-  line-height: 1.5;
-}
-
-.room-features {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 0.5rem;
-  margin-bottom: 1rem;
-}
-
-.feature {
-  background-color: #f0f3f5;
-  color: #2c3e50;
-  padding: 0.3rem 0.8rem;
-  border-radius: 20px;
-  font-size: 0.8rem;
-}
-
-.reserve-button {
-  background-color: #3498db;
-  color: white;
-  border: none;
-  padding: 1rem;
-  font-weight: 600;
-  cursor: pointer;
-  transition: background-color 0.3s;
-  margin-top: auto;
-}
-
-.reserve-button:hover {
-  background-color: #2980b9;
-}
-
-.reserve-button:disabled {
-  background-color: #bdc3c7;
-  cursor: not-allowed;
-}
