--- src/componentsAdmin/logout.css
@@ -1,56 +1,0 @@
-.logout-modal-overlay {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100vw;
-  height: 100vh;
-  background: rgba(0, 0, 0, 0.5); 
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  z-index: 1000;
-}
-
-.logout-modal {
-  background: white;
-  padding: 20px 30px;
-  border-radius: 8px;
-  box-shadow: 0 8px 16px rgba(0,0,0,0.3);
-  max-width: 300px;
-  text-align: center;
-}
-
-.logout-modal p {
-  font-size: 18px;
-  margin-bottom: 20px;
-  color: #333;
-}
-
-.logout-buttons {
-  display: flex;
-  justify-content: space-around;
-}
-
-.logout-buttons button {
-  background-color: #007bff; /* Azul */
-  border: none;
-  color: white;
-  padding: 10px 18px;
-  font-size: 16px;
-  border-radius: 5px;
-  cursor: pointer;
-  transition: background-color 0.3s ease;
-  min-width: 80px;
-}
-
-.logout-buttons button:hover {
-  background-color: #0056b3; 
-}
-
-.logout-buttons button:last-child {
-  background-color: #6c757d; 
-}
-
-.logout-buttons button:last-child:hover {
-  background-color: #5a6268; 
-}
