--- src/pagesAdmin/Suporte.js
@@ -1,12 +1,0 @@
-import React from 'react';
-
-function Suporte() {
-  return (
-    <div className='suporte'>
-      <h2>Suporte</h2>
-      <p>Bem vindo ao suporte mais rapido do brasil.</p>
-    </div>
-  );
-}
-
-export default Suporte;
